+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Don't count required keyword args when specifying `CountKeywordArgs: false` for `ParameterLists`. ([@sumeet][])
 * [#1879](https://github.com/bbatsov/rubocop/issues/1879): Avoid auto-correcting hash with trailing comma into invalid code in `BracesAroundHashParameters`. ([@jonas054][])
 * [#1868](https://github.com/bbatsov/rubocop/issues/1868): Do not register an offense in `Performance/Count` when `select` is called with symbols or strings as the parameters. ([@rrosenblum][])
+* `Sample` rewritten to properly handle shuffle randomness source, first/last params and non-literal ranges. ([@chastell][])
 
 ### New features
 
@@ -1413,3 +1414,4 @@
 [@l8nite]: https://github.com/l8nite
 [@sumeet]: https://github.com/sumeet
 [@ojab]: https://github.com/ojab
+[@chastell]: https://github.com/chastell